requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
moviepy==1.0.3
python-dotenv==0.20.0
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
urllib3==1.26.8
Werkzeug==2.0.2
google-cloud-texttospeech==2.10.0
flask-cors==3.0.10
beautifulsoup4==4.9.3
lxml==4.8.0
geopy==2.2.0
numpy==1.21.6
Pillow==9.1.0